+++ app/Http/Requests/Api/UpdateTaskListRequest.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace App\Http\Requests\Api;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class UpdateTaskListRequest extends FormRequest
+{
+    public function authorize()
+    {
+        return true;
+    }
+
+    public function rules()
+    {
+        return [
+            'name' => 'required|string|max:255',
+            'description' => 'nullable|string',
+            'tasks' => 'nullable|array',
+        ];
+    }
+}
